fix blobs

+++ examples/labels-3d.py
@@ -11,7 +11,7 @@ from napari.util import app_context
 
 with app_context():
     blobs = data.binary_blobs(length=128, volume_fraction=0.1, n_dim=3)
-    v = ViewerApp(blobs=blobs)
+    v = ViewerApp(blobs=blobs.astype(float))
     v.layers[0].colormap = 'gray'
     labeled = ndi.label(blobs)[0]
     label_layer = v.add_labels(labeled, name='blob ID')
